Unify LS searching logic whether it is scarb project or not.
Currently depending on if `Scarb.toml` was detected precedence of LS binary from settings ot from scarb is changing, remove this behaviour if favour of more stable one.
This is making usage of `cairo1.languageServerPath` and `cairo1.preferScarbLanguageServer` more predictable because it will work same way in scarb and non scarb projects.

commit-id:880ce024